refactor writeSize and readSize to utils because it is shared between node and web
