+++ ezmsg-js/src/web/bbuffer.ts
@@ -5,17 +5,8 @@ import {
   CreateNewBufferFunc,
   CreateBufferFromFunc
 } from '../shared/types';
-import {
-  BSIZE,
-  BITS_PER_BYTE,
-  SIZE_COUNT_BIT,
-  BYTE_1_WITH_SIZE_LIMIT,
-  BYTE_2_WITH_SIZE_LIMIT,
-  BYTE_1_WITH_SIZE_REST,
-  BYTE_2_WITH_SIZE_REST,
-  BYTE_4_WITH_SIZE_REST
-} from '../shared/constants';
-import { nextPowerOf2 } from '../shared/utils';
+import { BSIZE } from '../shared/constants';
+import { nextPowerOf2, writeSize, readSize } from '../shared/utils';
 
 const BFuncKey = [
   'Uint8',
@@ -67,36 +58,6 @@ export class BBuffer implements BBufferInterface {
     this.dv = newDV;
   }
 
-  writeSize(offset: number, size: number): number {
-    if (size < BYTE_1_WITH_SIZE_LIMIT) {
-      const writeVal = (1 << BYTE_1_WITH_SIZE_REST) | size;
-      return this.write(BType.U8, offset, writeVal);
-    }
-
-    if (size < BYTE_2_WITH_SIZE_LIMIT) {
-      const writeVal = (2 << BYTE_2_WITH_SIZE_REST) | size;
-      return this.write(BType.U16, offset, writeVal);
-    }
-
-    const writeVal = (4 << BYTE_4_WITH_SIZE_REST) | size;
-    return this.write(BType.U32, offset, writeVal);
-  }
-
-  readSize(offset): [number, number] {
-    const [firstByteValue] = this.read(BType.U8, offset);
-    const byteSize = (firstByteValue as number) >> BYTE_1_WITH_SIZE_REST;
-
-    const type =
-      byteSize === 1 ? BType.U8 : byteSize === 2 ? BType.U16 : BType.U32;
-
-    const [readVal, size] = this.read(type, offset);
-    const value =
-      (readVal as number) &
-      ((1 << (byteSize * BITS_PER_BYTE - SIZE_COUNT_BIT)) - 1);
-
-    return [value, size];
-  }
-
   write(type: BType, offset: number, value: BValue): number {
     if (type >= BType.U8 && type <= BType.F64) {
       this.expand(offset + BSIZE[type]);
@@ -112,7 +73,7 @@ export class BBuffer implements BBufferInterface {
       const encoded = encoder.encode(value as string);
 
       const valueSize = encoded.byteLength;
-      const typeSize = this.writeSize(offset, valueSize);
+      const typeSize = writeSize(this, offset, valueSize);
 
       this.expand(offset + valueSize + typeSize);
       this.arr.set(encoded, offset + typeSize);
@@ -127,7 +88,7 @@ export class BBuffer implements BBufferInterface {
     } else if (type === BType.BOOL) {
       return [this.dv[`get${BFuncKey[type]}`](offset) > 0, BSIZE[type]];
     } else if (type === BType.STR) {
-      const [valueSize, typeSize] = this.readSize(offset);
+      const [valueSize, typeSize] = readSize(this, offset);
       const dataOffset = offset + typeSize;
 
       const result = decoder.decode(
